modularized graph component
