+++ src/Components/Graphs.css
@@ -1,0 +1,7 @@
+.graphs-container {
+    position: fixed;
+    bottom: 5%;
+    right: 2.5%;
+    transform: translate(0, 0);
+    z-index: 1000;
+}
