+++ src/Components/Graphs.jsx
@@ -1,17 +1,33 @@
 //Component containing flowrate and current data
-const data = [30, 80, 45, 60, 20, 90, 50];
+import React, { useEffect, useState, useRef } from 'react';
+import generateGraph from '../api/MockData';
+import './Graphs.css'
 
-const svg = d3.select("#chart");
-const width = +svg.attr("width");
-const height = +svg.attr("height");
-const barWidth = width / data.length;
+function Graphs() {
+    const chartRef = useRef(null);
+    const [datainput, setDatainput] = useState(null);
 
-svg.selectAll("rect")
-.data(data)
-.enter()
-.append("rect")
-.attr("x", (_, i) => i * barWidth)
-.attr("y", d => height - d * 4) // Scale height
-.attr("width", barWidth - 5)
-.attr("height", d => d * 4)
-.attr("fill", "steelblue");
+    // Load JSON data that is being inputted from data-collection.py
+    useEffect(() => {
+    fetch('/data.json')
+        .then(res => res.json())
+        .then(data => {
+        setDatainput(data);
+        });
+    }, []);
+
+    // Render chart after data loads
+    useEffect(() => {
+    if (datainput) {
+        generateGraph(chartRef, datainput);
+    }
+    }, [datainput]);
+
+    return (
+        <div className="graphs-container">
+            <h1>Visual Graph of + component name</h1>
+            <svg ref = {chartRef}></svg>
+        </div>
+    )
+}
+export default Graphs;
