--- src/index.css
@@ -1,4 +1,0 @@
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
